BUGFIX use BigInt for OrderID
UPDATE provide task to reparse orders with ID `2147483647` due to use of `Int` DB Field

+++ src/Model/Order.php
@@ -42,7 +42,7 @@ class Order extends DataObject implements PermissionProvider
      */
     private static $db = [
         'StoreID' => 'Int',
-        'OrderID' => 'Int',
+        'OrderID' => 'BigInt',
         'Email' => 'Varchar(255)',
         'TransactionDate' => 'DBDatetime',
         'ProductTotal' => 'Currency',
